Submitting my personal dictionary
This is my personal dictionary, made since years.

This is just a way to let other people to check it and
cherry-pick what they want.

This is a mix of manual and automatic work.

Wrong words here are not in the main dict (it's an automatic
delete/resync).

There are also mistakes, of course. I usually find them
using it on various codebase.

I add in batch group words at the end of file. So, no sorting
now or in the future. That's to make it easier to read it
while time pass by.